backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
hypothesis>=6.98.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.24.0,<0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=42.0.0
